Refactor to use more local vars
Add a local "current_operand" to keep track of which operand we are
appending to.

Add a local "prev_event" to keep track of the button that was pressed
before. We use this variable to refine the behavior of the calculator to
make it match to the default calculator app on iOS.