Remove custom theme colors
Specifying colors in the `theme.colors` field of the
`tailwind.config.json` removes the default colors from Tailwind. Also I
think we should just stick to the default colors because it is easy to
reference using the Tailwind official documentation.

If we want to add additional colors we should use the field
`theme.extend.colors` inside the configuration file.